pyproject.toml
[project]
name = "portfolio"
version = "0.1.0"
description = "My own professional and personal Hub"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.3",
    "django-htmx>=1.27.0",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.41.0",
    "whitenoise>=6.12.0",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.2.0",
    "mkdocs-material>=9.7.4",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.12.0",
    "ruff>=0.15.5",
]

[tool.ruff]
target-version = "py314"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]  # pycodestyle, pyflakes, isort, pyupgrade

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.dev"
asyncio_mode = "auto"          # all async tests run without explicit mark
python_files = ["tests.py", "test_*.py"]

